Refactored 'codegen' dir to move proxy classes into 'codegen/proxy'. Also renamed 'codegen/utils' to 'codegen/util' #690
