+++ script/validators/source_validator.py
@@ -82,8 +82,8 @@ SKIP_FILES_LIST = [
     "src/include/index/bloom_filter.h",
     "src/include/index/compact_ints_key.h",
     "src/include/index/bwtree.h",
-    "src/codegen/utils/oa_hash_table.cpp",
-    "src/codegen/utils/cc_hash_table.cpp"
+    "src/codegen/util/oa_hash_table.cpp",
+    "src/codegen/util/cc_hash_table.cpp"
 ]
 
 ## ==============================================
