+++ src/codegen/hash_join_translator.cpp
@@ -13,7 +13,7 @@
 #include "codegen/hash_join_translator.h"
 
 #include "include/codegen/utils/if.h"
-#include "codegen/oa_hash_table_proxy.h"
+#include "include/codegen/proxy/oa_hash_table_proxy.h"
 #include "include/codegen/expression/tuple_value_translator.h"
 #include "codegen/vectorized_loop.h"
 #include "expression/tuple_value_expression.h"
