+++ src/codegen/oa_hash_table_proxy.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/oa_hash_table_proxy.h"
+#include "include/codegen/proxy/oa_hash_table_proxy.h"
 
 namespace peloton {
 namespace codegen {
