+++ src/codegen/operator/delete_translator.cpp
@@ -12,8 +12,8 @@
 
 #include "codegen/operator/delete_translator.h"
 
-#include "codegen/catalog_proxy.h"
-#include "codegen/transaction_runtime_proxy.h"
+#include "include/codegen/proxy/catalog_proxy.h"
+#include "include/codegen/proxy/transaction_runtime_proxy.h"
 #include "concurrency/transaction_manager_factory.h"
 
 namespace peloton {
