+++ src/codegen/operator/order_by_translator.cpp
@@ -13,8 +13,8 @@
 #include "codegen/operator/order_by_translator.h"
 
 #include "include/codegen/utils/if.h"
-#include "codegen/runtime_functions_proxy.h"
-#include "codegen/sorter_proxy.h"
+#include "include/codegen/proxy/runtime_functions_proxy.h"
+#include "include/codegen/proxy/sorter_proxy.h"
 #include "common/logger.h"
 #include "planner/order_by_plan.h"
