+++ src/codegen/operator/table_scan_translator.cpp
@@ -13,8 +13,8 @@
 #include "codegen/operator/table_scan_translator.h"
 
 #include "include/codegen/utils/if.h"
-#include "codegen/catalog_proxy.h"
-#include "codegen/transaction_runtime_proxy.h"
+#include "include/codegen/proxy/catalog_proxy.h"
+#include "include/codegen/proxy/transaction_runtime_proxy.h"
 #include "planner/seq_scan_plan.h"
 #include "storage/data_table.h"
