+++ src/codegen/tile_group.cpp
@@ -15,9 +15,9 @@
 #include "catalog/schema.h"
 #include "include/codegen/utils/if.h"
 #include "include/codegen/utils/loop.h"
-#include "codegen/runtime_functions_proxy.h"
+#include "include/codegen/proxy/runtime_functions_proxy.h"
 #include "codegen/scan_consumer.h"
-#include "codegen/tile_group_proxy.h"
+#include "include/codegen/proxy/tile_group_proxy.h"
 #include "codegen/type.h"
 #include "codegen/varlen.h"
 #include "codegen/vector.h"
