+++ src/codegen/type.cpp
@@ -13,7 +13,7 @@
 #include "codegen/type.h"
 
 #include "include/codegen/utils/if.h"
-#include "codegen/values_runtime_proxy.h"
+#include "include/codegen/proxy/values_runtime_proxy.h"
 #include "type/value.h"
 
 namespace peloton {
