+++ src/codegen/value.cpp
@@ -16,7 +16,7 @@
 #include <queue>
 
 #include "include/codegen/utils/if.h"
-#include "codegen/runtime_functions_proxy.h"
+#include "include/codegen/proxy/runtime_functions_proxy.h"
 #include "codegen/type.h"
 
 #include "type/value.h"
