+++ src/codegen/values_runtime_proxy.cpp
@@ -10,9 +10,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/values_runtime_proxy.h"
+#include "include/codegen/proxy/values_runtime_proxy.h"
 
-#include "codegen/value_proxy.h"
+#include "include/codegen/proxy/value_proxy.h"
 #include "type/value.h"
 
 namespace peloton {
