+++ src/include/common/macros.h
@@ -115,7 +115,7 @@ namespace peloton {
 #endif
 
 //===--------------------------------------------------------------------===//
-// utils
+// util
 //===--------------------------------------------------------------------===//
 
 #define ARRAY_NELEMS(a) (sizeof(a) / sizeof((a)[0]))
