+++ test/codegen/codegen_test_util.cpp
@@ -13,9 +13,9 @@
 #include "codegen/codegen_test_util.h"
 
 #include "type/value_factory.h"
-#include "codegen/runtime_functions_proxy.h"
-#include "codegen/values_runtime_proxy.h"
-#include "codegen/value_proxy.h"
+#include "include/codegen/proxy/runtime_functions_proxy.h"
+#include "include/codegen/proxy/values_runtime_proxy.h"
+#include "include/codegen/proxy/value_proxy.h"
 
 namespace peloton {
 namespace test {
