+++ test/codegen/function_builder_test.cpp
@@ -12,7 +12,7 @@
 
 #include "codegen/codegen.h"
 #include "codegen/function_builder.h"
-#include "codegen/runtime_functions_proxy.h"
+#include "include/codegen/proxy/runtime_functions_proxy.h"
 #include "common/harness.h"
 
 namespace peloton {
