+++ test/codegen/if_test.cpp
@@ -15,7 +15,7 @@
 #include "codegen/function_builder.h"
 #include "include/codegen/utils/if.h"
 #include "include/codegen/utils/loop.h"
-#include "codegen/runtime_functions_proxy.h"
+#include "include/codegen/proxy/runtime_functions_proxy.h"
 
 #include "codegen/codegen_test_util.h"
